feat: busqueda receta

+++ src/Restaurant/Ingrediente.java
@@ -14,7 +14,7 @@ public class Ingrediente extends DataInDB {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
 }
